src/main/java/com/syemon/usersystem/service/RequestValidator.java
package com.syemon.usersystem.service;

import jakarta.validation.ConstraintViolation;
import jakarta.validation.ConstraintViolationException;
import jakarta.validation.Validator;
import lombok.AllArgsConstructor;

import java.util.Set;
import java.util.stream.Collectors;

@AllArgsConstructor
class RequestValidator {
    public static final String DELIMITER = ", ";

    private final Validator validator;

    public <T> void validate(T request) {
        Set<ConstraintViolation<T>> violations = validator.validate(request);

        if (!violations.isEmpty()) {
            String message = violations.stream()
                    .map(ConstraintViolation::getMessage)
                    .sorted()
                    .collect(Collectors.joining(DELIMITER));

            throw new ConstraintViolationException("Error occurred: " + message, violations);
        }
    }

}
